requirements.txt
absl-py==0.7.1
aiofiles==0.4.0
aiohttp==3.5.4
aioresponses==0.5.2
alabaster==0.7.12
apixu==0.2.0
APScheduler==3.5.1
argh==0.26.2
asn1crypto==0.24.0
astor==0.7.1
async-generator==1.10
async-timeout==3.0.1
attrs==18.1.0
Automat==0.7.0
Babel==2.6.0
bleach==1.5.0
blis==0.2.4
boto3==1.5.20
botocore==1.8.50
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.5.2
colorama==0.4.1
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cookies==2.2.1
coverage==4.5.3
cryptography==2.6.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.9
docopt==0.6.2
docutils==0.14
duckling==1.8.0
en-core-web-md==2.0.0
en-core-web-sm==2.0.0
entrypoints==0.3
fakeredis==0.10.3
fbmessenger==5.2.0
Flask==1.0.2
Flask-Cors==3.0.4
Flask-JWT-Simple==0.0.3
freezegun==0.3.10
future==0.16.0
gast==0.2.2
gevent==1.2.2
graphviz==0.8.3
greenlet==0.4.15
grpcio==1.19.0
h5py==2.7.1
html5lib==0.9999999
httpretty==0.9.5
httptools==0.0.13
humanfriendly==4.18
hyperlink==17.3.1
idna==2.6
idna-ssl==1.1.0
imagesize==1.1.0
incremental==17.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
jmespath==0.9.4
JPype1==0.6.2
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter-core==4.4.0
kafka-python==1.4.4
Keras==2.1.6
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
klein==17.10.0
livereload==2.6.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==2.1.0
mattermostwrapper==2.1
mistune==0.8.4
mock==2.0.0
more-itertools==7.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
msgpack-python==0.5.4
multidict==4.5.2
murmurhash==0.28.0
nbconvert==5.4.1
nbformat==4.4.0
nbsphinx==0.3.2
networkx==2.1
numpy==1.14.3
packaging==17.1
pandas==0.24.2
pandocfilters==1.4.2
pathlib==1.0.1
pathtools==0.1.2
pbr==5.1.3
pika==0.12.0
plac==0.9.6
pluggy==0.6.0
port-for==0.3.1
preshed==1.0.1
prompt-toolkit==2.0.9
protobuf==3.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pykwalify==1.6.0
pymongo==3.7.2
pyOpenSSL==19.0.0
pypandoc==1.4
pyparsing==2.4.0
pyreadline==2.1
PySocks==1.6.8
pytest==3.5.1
pytest-cov==2.5.1
pytest-localserver==0.4.1
pytest-pycodestyle==1.3.1
pytest-sanic==0.1.12
python-crfsuite==0.9.6
python-dateutil==2.7.5
python-engineio==3.5.1
python-socketio==3.1.1
python-telegram-bot==10.1.0
pytz==2018.9
PyYAML==3.12
questionary==1.0.1
# Editable install with no version control (rasa-core==0.10.1)
-e d:\rasa_core-master
rasa-core-sdk==0.12.2
rasa-nlu==0.12.3
rasabaster==0.7.18
redis==2.10.6
regex==2017.4.5
requests==2.18.4
requests-toolbelt==0.9.1
responses==0.9.0
rocketchat-API==0.6.25
ruamel.yaml==0.15.37
s3transfer==0.1.13
sanic==18.12.0
Sanic-Cors==0.9.7
sanic-jwt==1.2.1
Sanic-Plugins-Framework==0.7.0
scikit-learn==0.19.1
scipy==1.1.0
service-identity==18.1.0
simplejson==3.13.2
six==1.11.0
sklearn-crfsuite==0.3.6
slackclient==1.2.1
snowballstemmer==1.2.1
spacy==2.0.11
Sphinx==1.8.2
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.2.5b1
sphinxcontrib-httpdomain==1.6.1
sphinxcontrib-programoutput==0.11
sphinxcontrib-trio==1.0.2
sphinxcontrib-versioning==2.2.1
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.18
srsly==0.0.5
tabulate==0.8.3
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminaltables==3.1.0
testpath==0.4.2
thinc==6.10.3
toolz==0.9.0
tornado==6.0.2
tqdm==4.23.3
traitlets==4.3.2
treq==17.8.0
twilio==6.14.0
Twisted==19.2.0
typing==3.6.4
typing-extensions==3.7.2
tzlocal==1.5.1
ujson==1.35
urllib3==1.22
wasabi==0.2.1
watchdog==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
webexteamssdk==1.1.1
websocket-client==0.54.0
websockets==6.0
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.10.11
yarl==1.3.0
zomatopy==1.0.10
zope.interface==4.6.0
